refactor(swarm): extract BaseNode and introduce node type hierarchy
- Extract shared infrastructure into BaseNode<N, B> generic struct
- Create node type hierarchy: BootNode, ClientNode, StorerNode
- Move builder infrastructure to dedicated builder.rs module
- Simplify TopologyBehaviour to use internal Command/Event types
- Clean up rustdoc across vertex-swarm-topology crate
- Remove separator comments and verbose documentation

The node hierarchy allows code reuse while supporting different
node configurations:
- BootNode: topology protocols only (handshake, hive, pingpong)
- ClientNode: topology + client protocols (pricing, retrieval, etc)
- StorerNode: client + storage protocols (storer, chunk sync)
