--- crates/swarm/node/src/node/behaviour.rs
@@ -1,99 +1,0 @@
-//! Network behaviour composition for SwarmNode.
-//!
-//! This module provides `NodeBehaviour` which combines:
-//! - `identify`: Peer identification protocol
-//! - `topology`: Handshake, hive, and pingpong protocols
-//! - `client`: Pricing, retrieval, and pushsync protocols
-//!
-//! The behaviour itself is non-generic because libp2p's derive macro
-//! doesn't easily support PhantomData. The generics are handled at the
-//! [`SwarmNode`] level instead.
-
-use std::sync::Arc;
-
-use crate::protocol::{BehaviourConfig as ClientBehaviourConfig, ClientBehaviour, ClientEvent};
-use libp2p::{identify, identity::PublicKey, swarm::NetworkBehaviour};
-use vertex_swarm_api::SwarmNodeTypes;
-use vertex_swarm_peermanager::AddressManager;
-use vertex_swarm_topology::{
-    BehaviourConfig as TopologyBehaviourConfig, SwarmTopologyBehaviour, TopologyEvent,
-};
-
-/// Combined network behaviour for a SwarmNode.
-///
-/// This composes the topology and client behaviours into a single swarm behaviour.
-/// Generic over `N: SwarmNodeTypes` to support different node configurations.
-#[derive(NetworkBehaviour)]
-#[behaviour(to_swarm = "NodeEvent")]
-pub struct NodeBehaviour<N: SwarmNodeTypes> {
-    /// Identify protocol - exchange peer info.
-    pub identify: identify::Behaviour,
-
-    /// Topology behaviour - handshake, hive, ping.
-    pub topology: SwarmTopologyBehaviour<N>,
-
-    /// Client behaviour - pricing, retrieval, pushsync.
-    pub client: ClientBehaviour,
-}
-
-impl<N: SwarmNodeTypes> NodeBehaviour<N> {
-    /// Create a new node behaviour.
-    pub fn new(local_public_key: PublicKey, identity: N::Identity) -> Self {
-        Self {
-            identify: identify::Behaviour::new(identify::Config::new(
-                "/vertex/1.0.0".to_string(),
-                local_public_key,
-            )),
-            topology: SwarmTopologyBehaviour::new(identity, TopologyBehaviourConfig::default()),
-            client: ClientBehaviour::new(ClientBehaviourConfig::default()),
-        }
-    }
-
-    /// Create a new node behaviour with address management.
-    pub fn with_address_manager(
-        local_public_key: PublicKey,
-        identity: N::Identity,
-        address_manager: Arc<AddressManager>,
-    ) -> Self {
-        Self {
-            identify: identify::Behaviour::new(identify::Config::new(
-                "/vertex/1.0.0".to_string(),
-                local_public_key,
-            )),
-            topology: SwarmTopologyBehaviour::with_address_manager(
-                identity,
-                TopologyBehaviourConfig::default(),
-                address_manager,
-            ),
-            client: ClientBehaviour::new(ClientBehaviourConfig::default()),
-        }
-    }
-}
-
-/// Events from the combined node behaviour.
-pub enum NodeEvent {
-    /// Identify protocol event.
-    Identify(identify::Event),
-    /// Topology event (peer ready, disconnected, discovered).
-    Topology(TopologyEvent),
-    /// Client event (pricing, retrieval, pushsync).
-    Client(ClientEvent),
-}
-
-impl From<identify::Event> for NodeEvent {
-    fn from(event: identify::Event) -> Self {
-        NodeEvent::Identify(event)
-    }
-}
-
-impl From<TopologyEvent> for NodeEvent {
-    fn from(event: TopologyEvent) -> Self {
-        NodeEvent::Topology(event)
-    }
-}
-
-impl From<ClientEvent> for NodeEvent {
-    fn from(event: ClientEvent) -> Self {
-        NodeEvent::Client(event)
-    }
-}
