+++ crates/swarm/topology/src/dns.rs
@@ -1,7 +1,7 @@
-//! Resolve `/dnsaddr/` multiaddrs to concrete addresses.
+//! DNS resolution for `/dnsaddr/` multiaddrs.
 //!
-//! libp2p's DNS transport only returns one TXT record for dnsaddr, but there may be many.
-//! This module resolves ALL records, enabling load distribution and failover.
+//! libp2p's DNS transport only returns one TXT record for dnsaddr. This module
+//! resolves ALL records, enabling load distribution and failover across bootnodes.
 
 use std::collections::HashSet;
 
@@ -11,7 +11,6 @@ use libp2p::Multiaddr;
 use libp2p::multiaddr::Protocol;
 use tracing::{debug, warn};
 
-/// Maximum recursion depth for DNS resolution to prevent infinite loops.
 const MAX_DNS_RECURSION_DEPTH: usize = 10;
 
 /// Errors from dnsaddr resolution.
@@ -20,18 +19,19 @@ pub enum DnsaddrResolveError {
     #[error("DNS lookup failed: {0}")]
     DnsLookup(String),
 
-    #[error("Maximum DNS recursion depth exceeded")]
+    #[error("maximum DNS recursion depth exceeded")]
     MaxRecursionDepth,
 }
 
 /// Resolve a `/dnsaddr/` multiaddr to concrete addresses.
-/// Non-dnsaddr inputs are returned as-is.
+///
+/// Non-dnsaddr inputs are returned unchanged.
 pub async fn resolve_dnsaddr(addr: &Multiaddr) -> Result<Vec<Multiaddr>, DnsaddrResolveError> {
     let mut seen = HashSet::new();
     resolve_dnsaddr_with_seen(addr, &mut seen).await
 }
 
-/// Resolve with a seen set to prevent loops when resolving multiple related addresses.
+/// Resolve with a seen set to prevent loops.
 pub async fn resolve_dnsaddr_with_seen(
     addr: &Multiaddr,
     seen: &mut HashSet<String>,
@@ -45,7 +45,6 @@ pub async fn resolve_all_dnsaddrs(addrs: impl IntoIterator<Item = &Multiaddr>) -
     let mut seen = HashSet::new();
 
     for addr in addrs {
-        // Skip non-dnsaddr addresses
         if !is_dnsaddr(addr) {
             resolved.push(addr.clone());
             continue;
@@ -53,16 +52,11 @@ pub async fn resolve_all_dnsaddrs(addrs: impl IntoIterator<Item = &Multiaddr>) -
 
         match resolve_dnsaddr_with_seen(addr, &mut seen).await {
             Ok(addrs) => {
-                debug!(
-                    addr = %addr,
-                    resolved_count = addrs.len(),
-                    "Resolved dnsaddr"
-                );
+                debug!(addr = %addr, resolved_count = addrs.len(), "Resolved dnsaddr");
                 resolved.extend(addrs);
             }
             Err(e) => {
                 warn!(addr = %addr, error = %e, "Failed to resolve dnsaddr");
-                // Include original - libp2p may be able to resolve it
                 resolved.push(addr.clone());
             }
         }
@@ -76,7 +70,7 @@ pub fn is_dnsaddr(addr: &Multiaddr) -> bool {
     addr.iter().any(|p| matches!(p, Protocol::Dnsaddr(_)))
 }
 
-/// Extract the domain from a `/dnsaddr/{domain}` multiaddr.
+/// Extract domain from `/dnsaddr/{domain}`.
 pub fn extract_dnsaddr_domain(addr: &Multiaddr) -> Option<String> {
     for proto in addr.iter() {
         if let Protocol::Dnsaddr(domain) = proto {
@@ -98,16 +92,11 @@ fn resolve_recursive<'a>(
             return Err(DnsaddrResolveError::MaxRecursionDepth);
         }
 
-        // Check if this is a dnsaddr
         let domain = match extract_dnsaddr_domain(addr) {
             Some(d) => d,
-            None => {
-                // Not a dnsaddr, return as-is
-                return Ok(vec![addr.clone()]);
-            }
+            None => return Ok(vec![addr.clone()]),
         };
 
-        // Prevent infinite loops from circular DNS references
         let cache_key = format!("_dnsaddr.{}", domain);
         if seen.contains(&cache_key) {
             debug!(domain = %domain, "Skipping already-seen dnsaddr domain");
@@ -115,10 +104,8 @@ fn resolve_recursive<'a>(
         }
         seen.insert(cache_key.clone());
 
-        // Create DNS resolver with system defaults
         let resolver = TokioResolver::tokio(ResolverConfig::default(), ResolverOpts::default());
 
-        // Query TXT records for _dnsaddr.{domain}
         let txt_name = format!("_dnsaddr.{}", domain);
         debug!(name = %txt_name, "Querying DNS TXT records");
 
@@ -132,22 +119,16 @@ fn resolve_recursive<'a>(
             for txt in record.txt_data() {
                 let txt_str = String::from_utf8_lossy(txt);
 
-                // Parse dnsaddr=<multiaddr> format
                 if let Some(value) = txt_str.strip_prefix("dnsaddr=") {
                     debug!(record = %value, "Found dnsaddr TXT record");
 
                     match value.parse::<Multiaddr>() {
                         Ok(resolved_addr) => {
-                            // Recursively resolve if this is also a dnsaddr
                             let nested = resolve_recursive(&resolved_addr, seen, depth + 1).await?;
                             results.extend(nested);
                         }
                         Err(e) => {
-                            warn!(
-                                value = %value,
-                                error = %e,
-                                "Failed to parse multiaddr from TXT record"
-                            );
+                            warn!(value = %value, error = %e, "Failed to parse multiaddr from TXT record");
                         }
                     }
                 }
@@ -176,7 +157,6 @@ mod tests {
 
     #[test]
     fn test_is_dnsaddr_false_for_dns4() {
-        // dns4 is NOT dnsaddr - it's a different protocol
         let addr: Multiaddr = "/dns4/example.com/tcp/1634".parse().unwrap();
         assert!(!is_dnsaddr(&addr));
     }
