+++ crates/swarm/topology/src/events.rs
@@ -1,7 +1,7 @@
-//! Events and commands for the topology behaviour.
+//! Public topology commands and events.
 //!
-//! All types use libp2p concepts (PeerId, Multiaddr). The client layer
-//! extracts Swarm-layer concepts (OverlayAddress) from HandshakeInfo/SwarmPeer.
+//! These are the external API types for interacting with [`TopologyBehaviour`](crate::TopologyBehaviour).
+//! Use [`TopologyCommand`] to send instructions and receive [`TopologyEvent`]s from the swarm.
 
 use libp2p::{Multiaddr, PeerId, swarm::ConnectionId};
 use vertex_net_handshake::HandshakeInfo;
@@ -10,17 +10,17 @@ use vertex_swarm_peer::SwarmPeer;
 /// Events emitted by the topology behaviour.
 #[derive(Clone)]
 pub enum TopologyEvent {
-    /// Peer completed handshake and is authenticated.
+    /// Peer completed handshake and is ready for communication.
     PeerAuthenticated {
         peer_id: PeerId,
         connection_id: ConnectionId,
         info: Box<HandshakeInfo>,
     },
 
-    /// All connections to a peer closed.
+    /// All connections to a peer have closed.
     PeerConnectionClosed { peer_id: PeerId },
 
-    /// Peer addresses received via hive.
+    /// Received peer addresses via hive protocol.
     HivePeersReceived { from: PeerId, peers: Vec<SwarmPeer> },
 
     /// Network depth changed.
@@ -30,7 +30,7 @@ pub enum TopologyEvent {
     DialFailed { address: Multiaddr, error: String },
 }
 
-/// Commands accepted by the topology behaviour.
+/// Commands for the topology behaviour.
 #[derive(Debug, Clone)]
 pub enum TopologyCommand {
     /// Dial a peer at the given address.
@@ -39,6 +39,6 @@ pub enum TopologyCommand {
     /// Close all connections to a peer.
     CloseConnection(PeerId),
 
-    /// Broadcast peer addresses via hive.
+    /// Broadcast peer addresses via hive protocol.
     BroadcastPeers { to: PeerId, peers: Vec<SwarmPeer> },
 }
