+++ crates/swarm/topology/src/lib.rs
@@ -1,20 +1,23 @@
 //! Swarm network topology management.
 //!
-//! This crate provides the libp2p behaviour and protocol handlers for Swarm topology.
-//! It handles peer discovery, bootnode DNS resolution, and topology events.
+//! Provides libp2p behaviour and handlers for Swarm peer discovery and connection
+//! management. Operates at the libp2p layer using `PeerId` and `Multiaddr`.
 //!
-//! # Abstraction Boundary
+//! # Public API vs Internal Types
 //!
-//! This crate operates at the libp2p layer and uses libp2p types (PeerId, Multiaddr,
-//! ConnectionId). The client layer (`vertex-client-peermanager`) handles the
-//! PeerId ↔ OverlayAddress mapping.
+//! This crate exposes two levels of command/event types:
+//!
+//! - **Public API** ([`TopologyCommand`], [`TopologyEvent`]): High-level commands and
+//!   events for the node layer. Use these to interact with the topology behaviour.
+//!
+//! - **Internal** (`handler::Command`, `handler::Event`): Low-level per-connection
+//!   messages between the behaviour and connection handlers. These are not exported.
 //!
 //! # Components
 //!
-//! - **Behaviour**: libp2p `NetworkBehaviour` for topology management
-//! - **Bootnode**: Bootstrap node connection management
-//! - **DNS**: Resolution of `/dnsaddr/` multiaddrs for bootnodes
-//! - **Events**: Topology commands and events using pure libp2p types
+//! - [`TopologyBehaviour`]: libp2p `NetworkBehaviour` managing handshake, hive, pingpong
+//! - [`BootnodeConnector`]: Bootstrap node connection strategy
+//! - [`dns`]: Resolution of `/dnsaddr/` multiaddrs
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
@@ -27,14 +30,13 @@ pub mod protocol;
 
 mod error;
 
-pub use behaviour::{Config as BehaviourConfig, SwarmTopologyBehaviour};
+pub use behaviour::TopologyBehaviour;
+pub use bootnode::BootnodeConnector;
 pub use dns::{DnsaddrResolveError, is_dnsaddr, resolve_all_dnsaddrs, resolve_dnsaddr};
 pub use error::{TopologyError, TopologyResult};
 pub use events::{TopologyCommand, TopologyEvent};
+pub use handler::TopologyConfig;
 pub use protocol::{
     TopologyInboundOutput, TopologyInboundUpgrade, TopologyOutboundInfo, TopologyOutboundOutput,
     TopologyOutboundRequest, TopologyOutboundUpgrade, TopologyUpgradeError,
 };
-
-// Re-export key types
-pub use bootnode::BootnodeConnector;
